remove the old, deprecated API
