+++ tsukuyomi/src/app/config.rs
@@ -1,13 +1,13 @@
 use {
     super::{
-        concurrency::{current_thread::CurrentThread, Concurrency, DefaultConcurrency},
-        path::{IntoPath, Path, RouteHandler},
+        concurrency::{Concurrency, DefaultConcurrency},
+        path::{IntoPath, Path, PathExtractor},
         recognizer::Recognizer,
         scope::{ScopeId, Scopes},
         App, AppInner, Endpoint, ScopeData, Uri,
     },
     crate::{
-        handler::{Handler, ModifyHandler},
+        handler::{metadata::Metadata, Handler, ModifyHandler},
         util::{Chain, Never},
     },
     std::{error, fmt, marker::PhantomData, rc::Rc, sync::Arc},
@@ -48,22 +48,6 @@ impl fmt::Display for Error {
 impl error::Error for Error {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         Some(&self.cause)
-    }
-}
-
-impl App {
-    /// Creates a new `App` from the provided configuration.
-    #[deprecated]
-    pub fn create(config: impl Config<()>) -> Result<Self> {
-        App::create_imp(config)
-    }
-}
-
-impl App<CurrentThread> {
-    /// Creates a new `App` from the provided configuration, without guarantees of thread safety.
-    #[deprecated]
-    pub fn create_local(config: impl Config<(), CurrentThread>) -> Result<Self> {
-        App::create_imp(config)
     }
 }
 
@@ -116,17 +100,6 @@ impl<'a, M, C> Scope<'a, M, C>
 where
     C: Concurrency,
 {
-    /// Adds a route onto the current scope.
-    #[deprecated]
-    pub fn route<H>(&mut self, handler: H) -> Result<()>
-    where
-        H: Handler,
-        M: ModifyHandler<H>,
-        M::Handler: Into<C::Handler>,
-    {
-        self.route2(handler)
-    }
-
     pub(crate) fn route2<H>(&mut self, handler: H) -> Result<()>
     where
         H: Handler,
@@ -164,53 +137,6 @@ where
         }
 
         Ok(())
-    }
-
-    /// Creates a sub-scope with the provided prefix onto the current scope.
-    #[deprecated]
-    pub fn mount(&mut self, prefix: impl AsRef<str>, config: impl Config<M, C>) -> Result<()> {
-        let prefix: Uri = prefix.as_ref().parse().map_err(Error::custom)?;
-
-        let scope_id = self
-            .scopes
-            .add_node(self.scope_id, {
-                let parent = &self.scopes[self.scope_id].data;
-                ScopeData {
-                    prefix: parent.prefix.join(&prefix).map_err(Error::custom)?,
-                    default_handler: None,
-                }
-            })
-            .map_err(Error::custom)?;
-
-        config
-            .configure(&mut Scope {
-                recognizer: &mut *self.recognizer,
-                scopes: &mut *self.scopes,
-                scope_id,
-                modifier: &*self.modifier,
-                _marker: PhantomData,
-            })
-            .map_err(Into::into)?;
-
-        Ok(())
-    }
-
-    /// Applies the specified configuration with a `ModifyHandler` on the current scope.
-    #[deprecated]
-    pub fn modify<M2>(
-        &mut self,
-        modifier: M2,
-        config: impl Config<Chain<&'a M, M2>, C>,
-    ) -> Result<()> {
-        config
-            .configure(&mut Scope {
-                recognizer: &mut *self.recognizer,
-                scopes: &mut *self.scopes,
-                scope_id: self.scope_id,
-                modifier: &Chain::new(self.modifier, modifier),
-                _marker: PhantomData,
-            })
-            .map_err(Into::into)
     }
 }
 
@@ -400,5 +326,116 @@ where
 
     fn configure(self, scope: &mut Scope<'_, M, C>) -> std::result::Result<(), Self::Error> {
         (self.0)(scope)
+    }
+}
+
+#[doc(hidden)]
+#[allow(missing_debug_implementations)]
+pub struct RouteHandler<E, T> {
+    endpoint: Arc<T>,
+    metadata: Metadata,
+    _marker: PhantomData<E>,
+}
+
+impl<E, T> RouteHandler<E, T>
+where
+    E: PathExtractor,
+    T: crate::endpoint::Endpoint<E::Output>,
+{
+    pub(crate) fn new(path: Path<E>, endpoint: T) -> Self {
+        let path = path.uri_str();
+        let endpoint = Arc::new(endpoint);
+
+        let mut metadata = match path {
+            "*" => Metadata::without_suffix(),
+            path => Metadata::new(path.parse().expect("this is a bug")),
+        };
+        *metadata.allowed_methods_mut() = endpoint.allowed_methods();
+
+        Self {
+            endpoint,
+            metadata,
+            _marker: PhantomData,
+        }
+    }
+}
+
+mod handle {
+    use {
+        super::{PathExtractor, RouteHandler},
+        crate::{
+            endpoint::{ApplyContext, Endpoint},
+            error::Error,
+            future::{Poll, TryFuture},
+            handler::{metadata::Metadata, Handler},
+            input::Input,
+        },
+        std::{marker::PhantomData, sync::Arc},
+    };
+
+    impl<E, T> Handler for RouteHandler<E, T>
+    where
+        E: PathExtractor,
+        T: Endpoint<E::Output>,
+    {
+        type Output = T::Output;
+        type Error = Error;
+        type Handle = RouteHandle<E, T>;
+
+        fn handle(&self) -> Self::Handle {
+            RouteHandle {
+                state: RouteHandleState::Init(self.endpoint.clone()),
+                _marker: PhantomData,
+            }
+        }
+
+        fn metadata(&self) -> Metadata {
+            self.metadata.clone()
+        }
+    }
+
+    #[doc(hidden)]
+    #[allow(missing_debug_implementations)]
+    pub struct RouteHandle<E, T>
+    where
+        E: PathExtractor,
+        T: Endpoint<E::Output>,
+    {
+        state: RouteHandleState<T, T::Future>,
+        _marker: PhantomData<E>,
+    }
+
+    #[allow(missing_debug_implementations)]
+    enum RouteHandleState<T, Fut> {
+        Init(Arc<T>),
+        InFlight(Fut),
+    }
+
+    impl<E, T> TryFuture for RouteHandle<E, T>
+    where
+        E: PathExtractor,
+        T: Endpoint<E::Output>,
+    {
+        type Ok = T::Output;
+        type Error = Error;
+
+        #[inline]
+        fn poll_ready(&mut self, input: &mut Input<'_>) -> Poll<Self::Ok, Self::Error> {
+            loop {
+                self.state = match self.state {
+                    RouteHandleState::Init(ref endpoint) => {
+                        let args = E::extract(input.params.as_ref())?;
+                        RouteHandleState::InFlight(
+                            endpoint
+                                .apply(args, &mut ApplyContext::new(input))
+                                .map_err(|(_args, err)| err)?,
+                        )
+                    }
+                    RouteHandleState::InFlight(ref mut in_flight) => {
+                        return in_flight.poll_ready(input).map_err(Into::into);
+                    }
+                };
+            }
+        }
     }
 }
