+++ tsukuyomi/src/lib.rs
@@ -18,7 +18,6 @@ mod generic;
 mod uri;
 
 pub mod app;
-pub mod config;
 pub mod endpoint;
 pub mod error;
 pub mod extractor;
